Continued setting up Competitors, got the Player working

+++ Writer Tycoon/Assets/Scripts/Entities/ICompetitor.cs
@@ -6,6 +6,6 @@ namespace WriterTycoon.Entities
 {
     public interface ICompetitor
     {
-        void IncreaseMoney(int amount);
+        void IncreaseMoney(float amount);
     }
 }
